behaviour of opening / closing submenus working
The behaviour is now right and I've also done a bit of performance stuff
(decreased dom renderings).
this should be still tested more and the positioning of submenus that
are nested very deep is still not considered and the may go off screen.
Also I might want to do a slight refactoring as the code has become
pretty spaghetti...

workstream: