fix #31 Position Calculation is wrong when body has css transform
When body has a transform applied, `position: fixed` behaves differently.
Fixed it by adding the scroll offset of the window to the viewport dimensions
and to the desired position

also fixed a bug where the submenu's 8px vertical offset was applied
to the root menu if it was attached to the dom via the `show` API
with an element as target.