fix #34 wrong position calculation
the size calculation was correct, but the position calculation was
off by a few pixels in edge-cases (literally 😅), because
`window.innerWidth` includes the scrollbar.
The 'clientWidth` property of the html element doesn't include the
scrollbars, which fixes the issue, but the html element does not
use 100% of the available height by default, thus the `min-height`
was added.

Also improved the css minify script a bit to work better with workbench
auto-formatting and added some documentation to it.